zturn: reference design: use MIO instead of EMIO for SD write protect
There is a MIO tied to GND (with 10K) with net label SDIO0_WP on the
zturn board. There is no need to use a block design and EMIO to
achieve the same result (SD not protected).

This simplifies the reference designe and make it easier to boot
without bitstream

Signed-off-by: Andrea Merello <andrea.merello@gmail.com>